+++ spec/policies/lottery_assignment_policy_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe LotteryAssignmentPolicy do
     let(:draw) { FactoryGirl.build_stubbed(:draw, status: 'lottery') }
     let(:lottery) { LotteryAssignment.new(draw: draw) }
 
-    permissions :index?, :create?, :update?, :automatic? do
+    permissions :index?, :create?, :update?, :automatic?, :export? do
       it { is_expected.not_to permit(user, lottery) }
     end
   end
@@ -22,7 +22,7 @@ RSpec.describe LotteryAssignmentPolicy do
       let(:draw) { FactoryGirl.build_stubbed(:draw, status: 'lottery') }
       let(:lottery) { LotteryAssignment.new(draw: draw) }
 
-      permissions :index?, :create?, :update? do
+      permissions :index?, :create?, :update?, :export? do
         it { is_expected.to permit(user, lottery) }
       end
 
@@ -34,7 +34,8 @@ RSpec.describe LotteryAssignmentPolicy do
       let(:draw) { FactoryGirl.build_stubbed(:draw, status: 'pre_lottery') }
       let(:lottery) { LotteryAssignment.new(draw: draw) }
 
-      permissions :index?, :create?, :update?, :automatic? do
+      permissions :index?, :create?, :update?, :automatic?,
+                  :export? do
         it { is_expected.not_to permit(user, lottery) }
       end
     end
@@ -43,7 +44,7 @@ RSpec.describe LotteryAssignmentPolicy do
   context 'admin' do
     let(:user) { FactoryGirl.build_stubbed(:user, role: 'admin') }
 
-    permissions :index?, :create?, :update?, :automatic? do
+    permissions :index?, :create?, :update?, :automatic?, :export? do
       context 'draw in lottery' do
         let(:draw) { FactoryGirl.build_stubbed(:draw, status: 'lottery') }
         let(:lottery) { LotteryAssignment.new(draw: draw) }